+++ src/events/RegisterFieldTransformersEvent.php
@@ -4,7 +4,7 @@ namespace samuelreichoer\queryapi\events;
 
 use craft\base\Event;
 
-class FieldTransformerEvent extends Event
+class RegisterFieldTransformersEvent extends Event
 {
     /**
      * @var array<int, array{fieldClass: string, transformer: class-string}>